+++ Customer/src/main/java/com/customer/model/PermanentAddress.java
@@ -15,7 +15,7 @@ public class PermanentAddress {
 	private String cityname;
 	private String district;
 	private String state;
-	private Long pincode;
+	private int pincode;
 	private int houseNumber;
 	private String streetName;
